+++ packages/react-schema-form/src/Field/createField.tsx
@@ -1,115 +1,15 @@
 import * as React from 'react';
-import {
-  Rule,
-  ValidationResult,
-  validate
-} from '@traveloka/validation';
-
-interface FieldInterface {
-  getValue: () => any,
-  setValue: (value: any) => any,
-  getError: () => ValidationResult,
-  setError: (error: ValidationResult) => ValidationResult,
-  validate: () => ValidationResult,
-  reset: () => void,
-};
-
-type FieldProps = {
-  revalidateOnError ?: boolean,
-  validateOnChange ?: boolean,
-  name?: string,
-  label?: string,
-  onChange?: (value:any) => void,
-  defaultValue?: any,
-  rules?: Rule,
-  [key: string]: any,
-}
-
-type FieldState = {
-  value: any,
-  error: ValidationResult,
-}
+import FieldComponent from '../Field';
 
 export default function createField(WrapperComponent: React.ComponentClass<any, any>) {
-  const Field = class extends React.Component<FieldProps, FieldState> implements FieldInterface {
-    public static defaultProps = {
-      revalidateOnError: true,
-      validateOnChange: false,
-    }
-
-    constructor(props: any) {
-      super(props);
-      this.state = {
-        value: props.defaultValue,
-        error: null,
-      }
-    }
-
+  const Field = class extends React.Component<any, any> {
     public render() {
-      const { name, ...rest } = this.props;
-      const label = this.ucwords(name);
       return (
-        <WrapperComponent
-          label={label}
-          {...rest}
-          name={name}
-          value={this.getValue()}
-          error={this.getError()}
-          onChange={this.handleOnChange}
+        <FieldComponent
+          component={WrapperComponent}
+          {...this.props}
         />
       );
-    }
-
-    public validate = () => {
-      const { rules = [] } = this.props;
-      const error = validate(rules)(this.getValue());
-      this.setError(error);
-      return error;
-    }
-
-    public getValue = () => {
-      return this.state.value;
-    }
-
-    public setValue = (value: any) => {
-      this.setState({
-        value
-      }, () => {
-        const { revalidateOnError, validateOnChange } = this.props;
-        if (validateOnChange || (revalidateOnError && this.getError())) {
-          this.validate();
-        }
-      });
-      if (this.props.onChange) {
-        this.props.onChange(value);
-      }
-    }
-
-    public getError = () => {
-      return this.state.error;
-    }
-
-    public setError = (error: ValidationResult) => {
-      this.setState({
-        error
-      });
-      return error;
-    }
-
-    public reset = () => {
-      this.setValue(null);
-      this.setError(null);
-    }
-
-    public handleOnChange = (value: any) => {
-      return this.setValue(value);
-    }
-
-    public ucwords = (str ?: string) => {
-      if (!str) return null;
-      return str.toLowerCase().replace(/\b[a-z]/g, letter => {
-        return letter.toUpperCase();
-      });
     }
   };
   return Field;
