+++ packages/react-schema-form/src/index.tsx
@@ -1,2 +1,4 @@
 export * from './types';
-export * from './createForm';
+export * from './Form/createForm';
+export {default as Form} from './Form';
+export {default as Field} from './Field';
