--- packages/react-schema-form/src/types/index.tsx
@@ -1,1 +1,0 @@
-export * from './Form';
